third_party/WebKit/LayoutTests/accessibility/aria-disabled-expected.txt

 
link  submit
treeitem
option
tab
gridcell
checkbox
This tests that the aria-disabled attribute works. The text field should not be enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textField.controlMode is "disabled"
PASS textField.controlMode is "none"
PASS group.controlMode is "disabled"
PASS link.controlMode is "disabled"
PASS button.controlMode is "none"
PASS treeitem.controlMode is "disabled"
PASS option.controlMode is "disabled"
PASS tab.controlMode is "disabled"
PASS gridcell.controlMode is "disabled"
PASS checkbox.controlMode is "disabled"
PASS successfullyParsed is true

TEST COMPLETE
